Batch-resolve private collections
## Ticket
1758

This alleviates the lock-up that is caused by large number of invidual resolves. There will still be some minor stutter due to the large DOM that React needs to handle -- that is logged in 1758 and will be handled separately.

At least the stutter is short (1-2s) and the app is still usable.

## Issue
Private list items are being resolve individually, super slow if the list is large (>100). Published lists doesn't have this issue.

## Assessment
doFetchItemsInCollections contains most of the useful logic, but it isn't called for private/built-in lists because it's not an actual claim.

## Approach
Tweaked doFetchItemsInCollections to handle private (UUID-based) collections.