PublishList: delay mounting Items tab to prevent lock-up
## Issue
For a large list, the playlist publish form is unusable (super-slow typing) due to the entire list being mounted despite the tab is not active.

## Approach
The full solution is still to paginate it, but for now, don't mount the tab until it is selected. Add a spinner to indicate something is loading. It's not prefect, but it's throwaway code anyway. At least we can fill in the fields properly now.